pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "criteria-bind"
version = "0.2.0"
description = "Hydra-driven Gemini reranker pipelines with MLflow tracking and offline demos."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
  {name = "Criteria Bind Team", email = "opensource@example.com"}
]
dependencies = [
  "torch>=2.3,<3.0",
  "transformers>=4.43,<5.0",
  "tokenizers>=0.19,<0.20",
  "datasets>=2.20,<3.0",
  "accelerate>=1.0.0,<2.0.0",
  "hydra-core>=1.3,<2.0",
  "omegaconf>=2.3,<3.0",
  "pydantic>=2.8,<3.0",
  "rich>=13.7,<14.0",
  "tqdm>=4.66,<5.0",
  "numpy>=1.24,<3.0",
  "pandas>=2.0,<3.0",
  "scikit-learn>=1.4,<2.0",
  "mlflow>=2.14,<3.0",
  "psutil>=5.9,<6.0",
  "tyro>=0.7.0,<1.0",
  "tenacity>=8.2,<9.0",
  "fastapi>=0.110,<1.0",
  "ujson>=5.9,<6.0",
  "sentencepiece>=0.1,<0.2",
  "nltk>=3.8,<4.0",
  "pyyaml>=6.0,<7.0",
  "regex>=2024.5.0",
  "python-dotenv>=1.0,<2.0",
  "evaluate>=0.4,<0.5",
  "pillow>=10.3,<11.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.2,<9.0",
  "pytest-cov>=5.0,<6.0",
  "ruff>=0.5.0,<0.6.0",
  "mypy>=1.10,<2.0",
  "types-requests>=2.32,<3.0",
  "types-setuptools>=69.0,<70.0",
  "pre-commit>=3.6,<4.0",
]
judge = [
  "google-genai>=0.6,<1.0",
  "uvicorn>=0.29,<0.30",
]

[project.scripts]
cb-train-criteria = "criteriabind.train.train_criteria_ranker:main"
cb-train-evidence = "criteriabind.train.train_evidence_span:main"
cb-judge = "criteriabind.cli.judge:main"
cb-pairs = "criteriabind.cli.pair_builder:main"
cb-infer = "criteriabind.cli.infer:main"
cb-eval = "criteriabind.cli.evaluate:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-q"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py310"
extend-select = ["I"]
extend-exclude = [
  "scripts",
  "src/criteriabind/config.py",
  "src/criteriabind/datasets/redsm5.py",
]

[tool.mypy]
python_version = "3.10"
warn_unused_configs = true
ignore_missing_imports = true
strict_optional = true
